ref: decoupled ddp2 (#3816)
